+++ component/componenttest/nop_exporter.go
@@ -20,7 +20,6 @@ import (
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/config"
 	"go.opentelemetry.io/collector/consumer/consumertest"
-	"go.opentelemetry.io/collector/internal/internalinterface"
 )
 
 // NewNopExporterCreateSettings returns a new nop settings for Create*Exporter functions.
@@ -35,54 +34,31 @@ type nopExporterConfig struct {
 	config.ExporterSettings `mapstructure:",squash"` // squash ensures fields are correctly decoded in embedded struct
 }
 
-// nopExporterFactory is factory for nopExporter.
-type nopExporterFactory struct {
-	internalinterface.BaseInternal
-}
-
-var nopExporterFactoryInstance = &nopExporterFactory{}
+var nopExporterFactory = component.NewExporterFactory(
+	"nop",
+	func() config.Exporter {
+		return &nopExporterConfig{
+			ExporterSettings: config.NewExporterSettings(config.NewComponentID("nop")),
+		}
+	},
+	component.WithTracesExporter(createTracesExporter),
+	component.WithMetricsExporter(createMetricsExporter),
+	component.WithLogsExporter(createLogsExporter))
 
 // NewNopExporterFactory returns a component.ExporterFactory that constructs nop exporters.
 func NewNopExporterFactory() component.ExporterFactory {
-	return nopExporterFactoryInstance
+	return nopExporterFactory
 }
 
-// Type gets the type of the Exporter config created by this factory.
-func (f *nopExporterFactory) Type() config.Type {
-	return "nop"
-}
-
-// CreateDefaultConfig creates the default configuration for the Exporter.
-func (f *nopExporterFactory) CreateDefaultConfig() config.Exporter {
-	return &nopExporterConfig{
-		ExporterSettings: config.NewExporterSettings(config.NewComponentID("nop")),
-	}
-}
-
-// CreateTracesExporter implements component.ExporterFactory interface.
-func (f *nopExporterFactory) CreateTracesExporter(
-	_ context.Context,
-	_ component.ExporterCreateSettings,
-	_ config.Exporter,
-) (component.TracesExporter, error) {
+func createTracesExporter(context.Context, component.ExporterCreateSettings, config.Exporter) (component.TracesExporter, error) {
 	return nopExporterInstance, nil
 }
 
-// CreateMetricsExporter implements component.ExporterFactory interface.
-func (f *nopExporterFactory) CreateMetricsExporter(
-	_ context.Context,
-	_ component.ExporterCreateSettings,
-	_ config.Exporter,
-) (component.MetricsExporter, error) {
+func createMetricsExporter(context.Context, component.ExporterCreateSettings, config.Exporter) (component.MetricsExporter, error) {
 	return nopExporterInstance, nil
 }
 
-// CreateLogsExporter implements component.ExporterFactory interface.
-func (f *nopExporterFactory) CreateLogsExporter(
-	_ context.Context,
-	_ component.ExporterCreateSettings,
-	_ config.Exporter,
-) (component.LogsExporter, error) {
+func createLogsExporter(context.Context, component.ExporterCreateSettings, config.Exporter) (component.LogsExporter, error) {
 	return nopExporterInstance, nil
 }